+++ tests/Unit/ExampleTest.php
@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Feature;
 
 use Orchestra\Testbench\TestCase;
 
-class ExampleTest extends TestCase
+final class ExampleTest extends TestCase
 {
     public function test_example(): void
     {